Adding IBO

+++ opengl/monitor.sh
@@ -1,7 +1,13 @@
 #!/bin/bash
-pt=$(ls -l --time-style=+%s circle.cpp|cut -d' ' -f6);
+fname=$1;
+echo "Monitoring file [$fname]";
+if [ ! -e $fname ] || [ "x" == "x$fname" ];then
+	echo "File fname not found"
+	exit 1
+fi
+pt=$(ls -l --time-style=+%s $fname|cut -d' ' -f6);
 while(/bin/true); do
-	ct=$(ls -l --time-style=+%s circle.cpp|cut -d' ' -f6);
+	ct=$(ls -l --time-style=+%s $fname|cut -d' ' -f6);
 	if [ "$ct" -ne "$pt" ];then
 		echo "************************************************";
 		killall triangle 2>&-;
